+++ app/models/__init__.py
@@ -5,3 +5,4 @@ from .equipe import Equipe
 from .circulo import Circulo
 from .encontreiro import Encontreiro
 from .encontrista import Encontrista
+from .detalhamento import Detalhamento